Remove trailing newline in code blocks before twoslash processing
markdown-it's parser produces code blocks with a trailing newline,
seemingly in conformance with the
[commonmark spec](https://spec.commonmark.org/0.12/#indented-code-blocks).
shiki-twoslash should avoid treating this line as contentful, as doing
so results in generated html with a "line" for that newline, adding an
extra newline to the user-provided code block. To avoid this unexpected
behavior, simply remove the trailing newline before twoslash-shiki
processing.

Closes #90